requirements.txt
# Core Flask Web Framework
Flask>=2.3.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0

# Forms and Validation
Flask-WTF>=1.1.0
WTForms>=3.0.0
email-validator>=2.0.0

# Machine Learning & NLP Dependencies
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0,<3.8.0

# Data Processing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0

# Configuration and Utilities
python-dotenv>=1.0.0

# Email Support
Flask-Mail>=0.9.0

# Production Server (optional)
gunicorn>=20.0.0

# Development Tools (optional)
# pytest>=7.0.0
# black>=22.0.0